Remove CAS1 Oasys API Generated Code
